refactor main hotkeys
